refactored unit tests to reduce code duplication
