+++ tests/unit_tests/commandline/test_geoips_config_create_registries.py
@@ -4,6 +4,10 @@ import pytest
 
 from geoips.commandline.commandline_interface import main
 from geoips.errors import PluginRegistryError
+from tests.unit_tests.commandline.cli_top_level_tester import (
+    generate_id,
+    check_valid_command_using_monkeypatch,
+)
 
 valid_args = [
     ["geoips", "config", "create-registries"],
@@ -64,22 +68,6 @@ invalid_args = [
 ]
 
 
-def generate_id(args):
-    """Generate an ID from the arguments provided.
-
-    Parameters
-    ----------
-    args: list[str]
-        - A list of strings representing the CLI command being ran.
-    expected: dict
-        - A dictionary of expected argument values that are returned from the command
-          being ran from args.
-    """
-    if isinstance(args, dict):
-        return ""
-    return " ".join(args)
-
-
 @pytest.mark.parametrize("cli_args", [args for args in invalid_args], ids=generate_id)
 def test_create_registries_invalid(monkeypatch, cli_args):
     """Test the CLI command 'geoips config create-registries' using invalid args.
@@ -93,6 +81,8 @@ def test_create_registries_invalid(monkeypatch, cli_args):
     if "-n" in cli_args or "-p" in cli_args:
         error = PluginRegistryError
     else:
+        # This occurs when 'save_type' is not one of 'json' or 'yaml'. Argparse
+        # eventually calls sys.exit(2, msg), which is caught here.
         error = SystemExit
     with pytest.raises(error):
         main()
@@ -114,10 +104,4 @@ def test_create_registries_valid(monkeypatch, cli_args, expected):
         - A dictionary of expected argument values that are returned from the command
           being ran from args.
     """
-    monkeypatch.setattr("sys.argv", cli_args)
-    args = main(suppress_args=False)
-    arg_dict = vars(args)
-    arg_dict.pop("command_parser")
-    arg_dict.pop("exe_command")
-    # Assert that the argument dictionary returned from the CLI matches what we expect
-    assert arg_dict == expected
+    check_valid_command_using_monkeypatch(monkeypatch, cli_args, expected)
